backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.6.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
firebase-admin==6.4.0
cloudinary==1.38.0
httpx==0.26.0
pydantic==2.7.4
pydantic-settings==2.3.0
python-dotenv==1.0.0
beanie==1.24.0
upstash-redis==1.1.0
slowapi==0.1.9
python-magic==0.4.27
bleach==6.1.0
